+++ packages/proteus/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.8.1](https://github.com/wireapp/wire-web-packages/tree/master/packages/proteus/compare/@wireapp/proteus@9.8.0...@wireapp/proteus@9.8.1) (2020-04-17)
+
+
+### Bug Fixes
+
+* **proteus:** Publish types for libsodium-wrappers-sumo ([b8a7fc6](https://github.com/wireapp/wire-web-packages/tree/master/packages/proteus/commit/b8a7fc6043225dcacd14d2c4e775d740ffd079c4))
+
+
+
+
+
+# [9.8.0](https://github.com/wireapp/wire-web-packages/tree/master/packages/proteus/compare/@wireapp/proteus@9.7.0...@wireapp/proteus@9.8.0) (2020-04-16)
+
+
+### Features
+
+* **proteus:** Use constructor instead of ClassUtil (BREAKING) (WEBAPP-6789) ([#2938](https://github.com/wireapp/wire-web-packages/tree/master/packages/proteus/issues/2938)) ([48e4c38](https://github.com/wireapp/wire-web-packages/tree/master/packages/proteus/commit/48e4c385cf819f42e3e7cefe566a2e49f32ba980))
+
+
+
+
+
 # [9.7.0](https://github.com/wireapp/wire-web-packages/tree/master/packages/proteus/compare/@wireapp/proteus@9.6.13...@wireapp/proteus@9.7.0) (2020-04-15)